requirements.txt
asgiref==3.5.2
Django==4.1.1
django-environ==0.8.1
pytz==2021.1
sqlparse==0.4.1
markdown
django-filter
celery
redis
requests
djangorestframework
markdown
django-filter